Anhadido valiadores para los pdfs

+++ node_js_api_mysql/server.js
@@ -907,14 +907,18 @@ app.get('/retrieve/:thisDocName', function(req, res) {
   
   const pathAxios = CARPETAPDF + "/" + thisDocName + '.pdf'; // path where to file is stored in server
 
-  //console.log("llegado al retrieve puro");
-  const rs = fs.createReadStream(pathAxios);
+  if (fs.existsSync(pathAxios)) {
+    //console.log("llegado al retrieve puro");
+    const rs = fs.createReadStream(pathAxios);
 
-  // get size of the video file
-  const { size } = fs.statSync(pathAxios);
-  res.setHeader("Content-Type", "application/pdf");
-  res.setHeader("Content-Length", size);
-  rs.pipe(res);
+    // get size of the video file
+    const { size } = fs.statSync(pathAxios);
+    res.setHeader("Content-Type", "application/pdf");
+    res.setHeader("Content-Length", size);
+    rs.pipe(res);
+  }else{
+    return res.status(500).json({ error: 'database connection error' });
+  }
   
 })
 
@@ -929,16 +933,18 @@ app.delete('/eliminate', function(req, res) {
   }
 
   console.log("id: " + id + " docName: " + docName);
-
   console.log("llegado al delete pdf para doc nombre " + docName);
 
   //------------------------------------------------
 
   console.log("llegado al delete pdf puro");
   const pathAxios = CARPETAPDF + "/" + docName  +".pdf";
-  // delete the file on server after it sends to client
-  const unlinkFile = util.promisify(fs.unlink); // to del file from local storage
-  unlinkFile(pathAxios);
+
+  if (fs.existsSync(pathAxios)) {
+    // delete the file on server after it sends to client
+    const unlinkFile = util.promisify(fs.unlink); // to del file from local storage
+    unlinkFile(pathAxios);
+  }
 
   //-----------------------------------------------
